refactoring
- board, score, bestScore, gameStatus 등 다양한 게임 플레이 관련된 요소를 원래 App.tsx에서 관리했으나, useGame.ts에서 관리하도록 옮김
- type 정의는 types.ts에, 상수는 constants.ts에 모아 둠
- 각 block의 색상 정보를 Block.tsx에 두었으나 이를 constants.ts로 이동
- blockColor type을 만들어 코드 길이를 줄이고 알아보기 쉽게 함
- MoveDirection type과 moveDirection 상수를 만들어 handleKeyDown에서의 동작을 switch 문 없이 단순화함
